shut down the delegate AmazonS3 instance
Previously, this root S3Decorator object did not have an implementation
for shutdown, so it calling shutdown on it would fall through to the
AbstractAmazonS3#shutdown implementation. That implementation just
throws an UnsupportedOperationException, which fails to shutdown any
underlying resources and often results in the calling program crashing.

Adding this shutdown override lets clients close the delegate's
underlying resources